cli: move options type definition
