+++ packages/cli/src/execute/serialize-output.ts
@@ -1,7 +1,7 @@
 import { writeFile } from 'node:fs/promises';
 import stringify from 'fast-safe-stringify';
 import { Logger } from '../util/logger';
-import { Opts } from '../commands';
+import { Opts } from '../options';
 const serializeOutput = async (
   options: Pick<Opts, 'strictOutput' | 'outputStdout' | 'outputPath'>,
   result: any,
